pyproject.toml
[tool.poetry]
name = "fastapi-tutorial"
version = "0.1.0"
description = ""
authors = ["jewon119 <jewon119@gmail.com>"]
readme = "README.md"
packages = [{include = "fastapi_tutorial"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.18"
psycopg2-binary = "^2.9.6"
pytest = "^7.4.0"
alembic = "^1.11.1"
python-decouple = "^3.8"
httpx = "^0.24.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
